--- python/chronos/example/nano_qp/tcn/tcn_config.py
@@ -1,14 +1,0 @@
-lookback = 48 # number of steps to look back
-horizon = 1 # number of steps to predict
-input_feature_num = 8 # number of feature to use
-output_feature_num = horizon
-past_seq_len = lookback
-feature_seq_len = 1 # number of feature to predict
-optimizer = 'Adam'
-lr = 0.001
-batch_size = 32
-num_epochs = 3
-seed = 0
-device = "gpu"
-save_dir = './model'
-save_model_file = f'horizon:{horizon}lookback:{lookback}lr:{lr}_epoch:{num_epochs}' + 'tcn.pth'
